pyproject.toml
[project]
name = "hig"
version = "0.1.0"
description = "Vietnamese Historical Image Generation with Flux.1"
readme = "README.md"
authors = [
    { name = "TuThanh", email = "tuthanh10825@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.26.0",
    "bitsandbytes>=0.41.0",
    "datasets>=2.16.0",
    "diffusers>=0.27.0",
    "gradio>=4.0.0",
    "huggingface-hub>=0.20.0",
    "llama-cpp-python>=0.2.0",
    "peft>=0.8.0",
    "pillow>=10.0.0",
    "sentencepiece>=0.1.99",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "tqdm>=4.65.0",
    "transformers>=4.37.0",
]

[project.scripts]
hig = "hig:main"

[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
